- Changed: GameObject.Realm has now changed from a byte to an eRealm (which can still be cast to a byte).

+++ GameServer/gameobjects/GameTeleporter.cs
@@ -85,7 +85,7 @@ namespace DOL.GS
 				{
 					Teleport teleport = new Teleport();
 					teleport.TeleportID = "battlegrounds";
-					teleport.Realm = portalKeep.Realm;
+					teleport.Realm = (byte)portalKeep.Realm;
 					teleport.RegionID = portalKeep.Region;
 					teleport.X = portalKeep.X;
 					teleport.Y = portalKeep.Y;
